compilation fails on CUDA 9 because of mandatory compute_20
Using Ubuntu 16.04 with a GTX 1060.

If you install CUDA today from https://developer.nvidia.com/cuda-downloads the version you get by default is 9.1. If you try to compile ccminer with this CUDA version, it fails. The reason is that compute_20 has been removed from CUDA 9 (it was deprecated already for a while in previous versions), and the ccminer source contains compute_20 directives in a few places.

The changes I propose in this PR are most certainly wrong. **Please do no merge**. But I cannot open an issue with your Github project, so this seems the best way to communicate.

Anyway, I've compiled my copy of ccminer with the changes shown above and now it's mining lyra2v2 just fine on a GTX 1060.

Something needs to be done about compute_20 and CUDA 9. Either remove compute_20 or add some conditional to skip it on the current version of CUDA.